Begin extracting JsAudioPlayer
